backend/foodgram/requirements.txt
Django==3.2
djangorestframework==3.12.4
django-filter~=22.1
drf-extra-fields==3.4.1
Pillow==9.5.0
gunicorn==20.0.4
psycopg2-binary==2.9.6
python-dotenv==0.19.0
